Preact: Support backing up teams
(Restoring backups not yet implemented)
