+++ play.pokemonshowdown.com/src/battle-teams.ts
@@ -1,5 +1,5 @@
-import { Dex, type ModdedDex } from "./battle-dex";
-import { BattleNatures, BattleStatNames } from "./battle-dex-data";
+import { Dex, toID, type ModdedDex } from "./battle-dex";
+import { BattleNatures, BattleStatNames, BattleStatIDs, type StatNameExceptHP, type ID } from "./battle-dex-data";
 
 export declare namespace Teams {
 	/**
@@ -46,6 +46,14 @@ export declare namespace Teams {
 		species: string;
 		moves: string[];
 	}
+	export interface Team {
+		name: string;
+		format: ID;
+		folder: string;
+		/** Note that this can be wrong if `.uploaded?.notLoaded` */
+		packedTeam: string;
+		isBox: boolean;
+	}
 }
 
 export const Teams = new class {
@@ -85,18 +93,10 @@ export const Teams = new class {
 				evs = `|${set.evs['hp'] || ''},${set.evs['atk'] || ''},${set.evs['def'] || ''},` +
 					`${set.evs['spa'] || ''},${set.evs['spd'] || ''},${set.evs['spe'] || ''}`;
 			}
-			if (evs === '|,,,,,') {
-				buf += '|';
-			} else {
-				buf += evs;
-			}
+			buf += evs === '|,,,,,' ? '|' : evs;
 
 			// gender
-			if (set.gender) {
-				buf += `|${set.gender}`;
-			} else {
-				buf += '|';
-			}
+			buf += `|${set.gender || ''}`;
 
 			// ivs
 			let ivs = '|';
@@ -104,32 +104,16 @@ export const Teams = new class {
 				ivs = `|${getIv(set.ivs, 'hp')},${getIv(set.ivs, 'atk')},${getIv(set.ivs, 'def')},` +
 					`${getIv(set.ivs, 'spa')},${getIv(set.ivs, 'spd')},${getIv(set.ivs, 'spe')}`;
 			}
-			if (ivs === '|,,,,,') {
-				buf += '|';
-			} else {
-				buf += ivs;
-			}
+			buf += ivs === '|,,,,,' ? '|' : ivs;
 
 			// shiny
-			if (set.shiny) {
-				buf += '|S';
-			} else {
-				buf += '|';
-			}
+			buf += `|${set.shiny ? 'S' : ''}`;
 
 			// level
-			if (set.level && set.level !== 100) {
-				buf += `|${set.level}`;
-			} else {
-				buf += '|';
-			}
+			buf += `|${set.level && set.level !== 100 ? set.level : ''}`;
 
 			// happiness
-			if (set.happiness !== undefined && set.happiness !== 255) {
-				buf += `|${set.happiness}`;
-			} else {
-				buf += '|';
-			}
+			buf += `|${set.happiness !== undefined && set.happiness !== 255 ? set.happiness : ''}`;
 
 			if (set.pokeball || set.hpType || set.gigantamax ||
 				(set.dynamaxLevel !== undefined && set.dynamaxLevel !== 10) || set.teraType) {
@@ -151,6 +135,16 @@ export const Teams = new class {
 
 	unpack(buf: string): Teams.PokemonSet[] {
 		if (!buf) return [];
+
+		// first, detect if this has team metadata
+		const endIndex = buf.indexOf(']');
+		if (endIndex > 0) {
+			const firstPart = buf.slice(0, endIndex);
+			const pipeCount = firstPart.split('|').length - 1;
+			if (pipeCount === 12 || pipeCount === 1) {
+				buf = buf.slice(buf.indexOf('|') + 1);
+			}
+		}
 
 		const team = [];
 		let i = 0;
@@ -426,5 +420,162 @@ export const Teams = new class {
 			text += Teams.exportSet(set, dex, newFormat);
 		}
 		return text;
+	}
+
+	parseExportedTeamLine(line: string, isFirstLine: boolean, set: Dex.PokemonSet) {
+		if (isFirstLine || line.startsWith('[')) {
+			let item;
+			[line, item] = line.split('@');
+			line = line.trim();
+			item = item?.trim();
+			if (item) {
+				set.item = item;
+				if (toID(set.item) === 'noitem') set.item = '';
+			}
+			if (line.endsWith(' (M)')) {
+				set.gender = 'M';
+				line = line.slice(0, -4);
+			}
+			if (line.endsWith(' (F)')) {
+				set.gender = 'F';
+				line = line.slice(0, -4);
+			}
+			if (line.startsWith('[') && line.endsWith(']')) {
+				// the ending `]` is necessary to establish this as ability
+				// (rather than nickname starting with `[`)
+				set.ability = line.slice(1, -1);
+				if (toID(set.ability) === 'selectability') {
+					set.ability = '';
+				}
+			} else if (line) {
+				const parenIndex = line.lastIndexOf(' (');
+				if (line.endsWith(')') && parenIndex !== -1) {
+					set.species = Dex.species.get(line.slice(parenIndex + 2, -1)).name;
+					set.name = line.slice(0, parenIndex);
+				} else {
+					set.species = Dex.species.get(line).name;
+					set.name = '';
+				}
+			}
+		} else if (line.startsWith('Trait: ')) {
+			set.ability = line.slice(7);
+		} else if (line.startsWith('Ability: ')) {
+			set.ability = line.slice(9);
+		} else if (line.startsWith('Item: ')) {
+			set.item = line.slice(6);
+		} else if (line.startsWith('Nickname: ')) {
+			set.name = line.slice(10);
+		} else if (line.startsWith('Species: ')) {
+			set.species = line.slice(9);
+		} else if (line === 'Shiny: Yes' || line === 'Shiny') {
+			set.shiny = true;
+		} else if (line.startsWith('Level: ')) {
+			set.level = +line.slice(7);
+		} else if (line.startsWith('Happiness: ')) {
+			set.happiness = +line.slice(11);
+		} else if (line.startsWith('Pokeball: ')) {
+			set.pokeball = line.slice(10);
+		} else if (line.startsWith('Hidden Power: ')) {
+			set.hpType = line.slice(14);
+		} else if (line.startsWith('Dynamax Level: ')) {
+			set.dynamaxLevel = +line.slice(15);
+		} else if (line === 'Gigantamax: Yes' || line === 'Gigantamax') {
+			set.gigantamax = true;
+		} else if (line.startsWith('Tera Type: ')) {
+			set.teraType = line.slice(11);
+		} else if (line.startsWith('EVs: ')) {
+			const evLines = line.slice(5).split('(')[0].split('/');
+			set.evs = { hp: 0, atk: 0, def: 0, spa: 0, spd: 0, spe: 0 };
+			let plus = '', minus = '';
+			for (let evLine of evLines) {
+				evLine = evLine.trim();
+				const spaceIndex = evLine.indexOf(' ');
+				if (spaceIndex === -1) continue;
+				const statid = BattleStatIDs[evLine.slice(spaceIndex + 1)];
+				if (!statid) continue;
+				if (evLine.charAt(spaceIndex - 1) === '+') plus = statid;
+				if (evLine.charAt(spaceIndex - 1) === '-') minus = statid;
+				set.evs[statid] = parseInt(evLine.slice(0, spaceIndex), 10) || 0;
+			}
+			const nature = this.getNatureFromPlusMinus(plus as StatNameExceptHP, minus as StatNameExceptHP);
+			if (nature) set.nature = nature;
+		} else if (line.startsWith('IVs: ')) {
+			const ivLines = line.slice(5).split(' / ');
+			set.ivs = { hp: 31, atk: 31, def: 31, spa: 31, spd: 31, spe: 31 };
+			for (let ivLine of ivLines) {
+				ivLine = ivLine.trim();
+				const spaceIndex = ivLine.indexOf(' ');
+				if (spaceIndex === -1) continue;
+				const statid = BattleStatIDs[ivLine.slice(spaceIndex + 1)];
+				if (!statid) continue;
+				let statval = parseInt(ivLine.slice(0, spaceIndex), 10);
+				if (isNaN(statval)) statval = 31;
+				set.ivs[statid] = statval;
+			}
+		} else if (/^[A-Za-z]+ (N|n)ature/.exec(line)) {
+			let natureIndex = line.indexOf(' Nature');
+			if (natureIndex === -1) natureIndex = line.indexOf(' nature');
+			if (natureIndex === -1) return;
+			line = line.slice(0, natureIndex);
+			if (line !== 'undefined') set.nature = line as Dex.NatureName;
+		} else if (line.startsWith('-') || line.startsWith('~') || line.startsWith('Move:')) {
+			if (line.startsWith('Move:')) line = line.slice(4);
+			line = line.slice(line.charAt(1) === ' ' ? 2 : 1);
+			if (line.startsWith('Hidden Power [')) {
+				let hpType = line.slice(14, line.indexOf(']')) as Dex.TypeName;
+				if (hpType.includes(']') || hpType.includes('[')) hpType = '' as any;
+				line = 'Hidden Power ' + hpType;
+				set.hpType = hpType;
+			}
+			if (line === 'Frustration' && set.happiness === undefined) {
+				set.happiness = 0;
+			}
+			set.moves.push(line);
+		}
+	}
+	getNatureFromPlusMinus(
+		plus: StatNameExceptHP | '' | null, minus: StatNameExceptHP | '' | null
+	): Dex.NatureName | null {
+		if (!plus || !minus) return null;
+		for (const i in BattleNatures) {
+			if (BattleNatures[i as 'Serious'].plus === plus && BattleNatures[i as 'Serious'].minus === minus) {
+				return i as Dex.NatureName;
+			}
+		}
+		return null;
+	}
+	import(buffer: string): Dex.PokemonSet[] {
+		const lines = buffer.split("\n");
+
+		const sets: Dex.PokemonSet[] = [];
+		let curSet: Dex.PokemonSet | null = null;
+
+		while (lines.length && !lines[0]) lines.shift();
+		while (lines.length && !lines[lines.length - 1]) lines.pop();
+
+		if (lines.length === 1 && lines[0].includes('|')) {
+			return Teams.unpack(lines[0]);
+		}
+		for (let line of lines) {
+			line = line.trim();
+			if (line === '' || line === '---') {
+				curSet = null;
+			} else if (line.startsWith('===')) {
+				// team backup format; ignore
+			} else if (line.includes('|')) {
+				// packed format
+				return Teams.unpack(line);
+			} else if (!curSet) {
+				curSet = {
+					name: '', species: '', gender: '',
+					moves: [],
+				};
+				sets.push(curSet);
+				this.parseExportedTeamLine(line, true, curSet);
+			} else {
+				this.parseExportedTeamLine(line, false, curSet);
+			}
+		}
+		return sets;
 	}
 };
